+++ pkg/textual/io_reader_processor.go
@@ -19,8 +19,6 @@ import (
 	"context"
 	"io"
 	"time"
-
-	"github.com/benoit-pereira-da-silva/textual/pkg/carrier"
 )
 
 // IOReaderProcessor connects an io.Reader to a Processor by scanning the input
@@ -54,7 +52,7 @@ import (
 //
 // Note: methods such as FromUTF8String are typically called on the zero value
 // of S, so implementations must not depend on receiver state.
-type IOReaderProcessor[S carrier.Carrier[S], P Processor[S]] struct {
+type IOReaderProcessor[S Carrier[S], P Processor[S]] struct {
 	reader    io.Reader
 	splitFunc bufio.SplitFunc // splitFunc defines the bufio.SplitFunc used to tokenize the input from the io.Reader.
 	processor P
@@ -69,7 +67,7 @@ type IOReaderProcessor[S carrier.Carrier[S], P Processor[S]] struct {
 // NewIOReaderProcessor constructs a new IOReaderProcessor using the provided
 // processor and reader. By default, it uses bufio.ScanLines as a split function
 // and a background context created on the first Start / StartWithTimeout.
-func NewIOReaderProcessor[S carrier.Carrier[S], P Processor[S]](processor P, reader io.Reader) *IOReaderProcessor[S, P] {
+func NewIOReaderProcessor[S Carrier[S], P Processor[S]](processor P, reader io.Reader) *IOReaderProcessor[S, P] {
 	return &IOReaderProcessor[S, P]{
 		splitFunc: bufio.ScanLines,
 		reader:    reader,